[BUGFIX] Only flush() the Doctrine ORM if it is open
In some cases it can happen, that the EntityManager is closed at the end
of a request This can be a valid situation (see #30933 for more).

This change makes persistAll() check if the EntityManager is open
before calling flush() on it.

Change-Id: I529509b84032932cb3f4db1dad3cd8bcb37e5e89
Fixes: #30934
Releases: 1.0, 1.1
Original-Commit-Hash: 491119035d378be54da1894ee083d1f5dacc5dde