[FEATURE] Add __toString() and make Resource use more convenient
To use a Resource with file operations, one can use the
resource:// scheme and the __toString() method of any
ResourcePointer.

To make this a little easier, this change adds two new
methods to the Resource class:

__toString() returns the SHA1 of the ResourcePointer the
Resource is holding (as is documented for ages).

getUri() returns a ready-to-use resource://<sha1> URI of
the ResourcePointer the Resource is holding.

Change-Id: Iec342eae0662b89eaf1d281fcab26ebb94ef7c61
Fixes: #29694
Resolves: #34177
Releases: 1.0, 1.1
Original-Commit-Hash: 8c7ec51b5a0e90a859c102fde6f5f0859babed1a