[FEATURE] Make HTTP responses standards compliant
This adds a new API method to the HTTP response which
is called by a request handler to assure conformity
with RFC 2616 and other related specifications.

While there are a lot of rules to consider, this patch
only introduces a first set of important checks which
are all explained by a corresponding test.

Change-Id: I8b7a9bf53061cfc371c22bdec018c1918c63a4fe
Related: #33371
Releases: 1.1, 1.2
Original-Commit-Hash: 462fec11312d0ad520e0ced8a201bf7adde2e544