[BUGFIX] Non-standard ports are now considered by HttpRequest
Ports that are not 80 for HTTP or 443 for HTTPS were simply not
passed through to the Request when being created by the create()
method.

Also Uri will now set the port to 443 for HTTPS if no port is
given in the string handed to the constructor.

Change-Id: I52be3aa3d04cb9407c663a5df4501e6ed04c9995
Fixes: #37770
Releases: 1.1, 1.2
Original-Commit-Hash: 091a37f7ab8332fe69b79df5659bf1559235099a