[BUGFIX] Check inheritance in TypeHandling::isCollectionType()
With this change TypeHandling::isCollectionType() does consider
inheritance when it comes to the checking it does. Thus a subclass of
SplObjectStorage or an implementation of Doctrine's Collection interface
will now be seen as a collection type.

Change-Id: I007b0faf486680ecfb2fe7c753ef4cabcf6d618c
Fixes: #44535
Releases: master, 2.0, 1.1
Original-Commit-Hash: 59c0976870264f59ca88b186f27b2a89c4c2cd4e