[BUGFIX] Throw exception when using options with singleton validators
Validators being singleton must not be called with options because
the options are then inherent properties of the Validator and cannot
be executed subsequently with other options.

An exception is thrown if such a case occurs because it is then
a design mistake of the developer.

Besides, this improves the namespace use a bit.

Change-Id: I4a7dbaaaf8fa39a3f4677bb43db044a3231b843f
Releases: master, 2.0
Fixes: #34872
Original-Commit-Hash: 05fc389ba3bf7022960cfd6cd9f66f1f82732f11