[TASK] Throw exception on lacking @var annotation
In the Validation\ValidationResolver now an
exception is thrown if the validated
property does not have an @var annotation
which is mandatory.

This prevents a Notice exception when
accessing the not present 'var' key
on the tag values array, and gives instead
a more helpful exception message.

Releases: master, 2.0
Change-Id: I293006d1bbe9d069e9f64aff57ae47a243009927
Original-Commit-Hash: d5fff3f64b54437db774e922b85262abac4e2723