[BUGFIX] ReflectionService works for non-Entity inheritance members
As soon as a class which is *not* an entity was present as an
ancestor of an Entity (with at least one Entity descendant in
between), there was an Undefined Index error because the
ReflectionService proceeded on the assumption that everything
was an Entity, hence present in the ->classSchemata array.

This adds an additional check for the actual presence of
such an array index and proofs this behavior with a
Functional Test Fixture.

Fixes: #47809
Releases: master, 2.0
Change-Id: I12118004e60e2230a7021a73251ea7c57d660d3b
Original-Commit-Hash: be7b184a07fff9b4b1f15fe6d189a34abf5be5e6