[BUGFIX] Fix resolving of imported collection types
This change extends the ReflectionService so that it is able to
expand our custom notation for collection types::

 CollectionType<ElementType>

Background:

In the Coding Guidelines we encourage developers to import namespaces
in order to increase readability.

With #46008 the ReflectionService is able to resolve relative and
imported namespace (with #50909 this also works for abstract classes).

But this does not work for our custom notation of Collection types yet.
With this fix relative and imported class names are supported also for
our custom notation of collection types and the above would be
expanded to::

 \CollectionTypeNamespace\CollectionType<\ElementTypeNamespace\ElementType>

Change-Id: I4e228d6846776d2500d7e8aec2bc88df170bc9de
Fixes: #57034
Related: #46008
Related: #50909
Releases: master, 2.2, 2.1
Original-Commit-Hash: 0b45aa53d72a61241d975d5d68a64ea638439306