[BUGFIX] Requesting /index.phpa/ breaks Uri host parameter
If the request URI contains ``index.php`` and something afterwards
before a slash, that something is included in the host being
set on the request URI that's constructed.

Problem lies in the handling of URIs containing index.php,
where that part is automatically stripped off. However it
strips it off in a way that can cause this issue, when the
``parse_url()`` is called on a broken URI.

Change-Id: Ib7fcbbfcaed3d8c2cbb437eb1a7796b7c1edd4f7
Releases: master, 2.2, 2.1, 2.0
Fixes: FLOW-92
Original-Commit-Hash: 33833deb4c4b98778ccb9fa88656fbeeb8500009