[BUGFIX] Disable security for CLI requests
Currently it's not possible to invoke methods that are covered by a
policy via CLI because the security context is not yet initialized
leading to a ``The security Context cannot be initialized yet``
exception.

With this change all authorization checks are disabled for command
controllers.

Change-Id: I1dc9d463bbe651f5cfc4faa646b35ae8bac85fe5
Releases: master, 2.3, 2.2, 2.1
Fixes: #FLOW-163
Original-Commit-Hash: fb4fe0d6590aa2e06e93aef375dfbea74bae0549