BUGFIX: No longer pick up too much YAML in migration
The migration to adjust Role persistence reads YAML files to learn
about existing roles. It picked up "too much" sometimes, which would
lead to errors, e.g. if some YAML could not be parsed but was not at
all relevant for the task.

This change makes the check on the path of found files stricter to
avoid that.