BUGFIX: Set orphanRemoval for non-AggregateRoots
After fixing the issue #1079, orphan removal is per default false for properties from classes which are no aggregate roots,
because the default value for orphan removal in the ManyToMany, OneToOne and OneToMany annotations is "false".
Unfortunately we can not prevent the default value in the annotation, so we can not rely on the property not being set to fall
back to our defaults.
This change fixes that by switching the check for orphanRemoval annotation value fallbacks and always apply our default value
for non-aggregate root non-value objects.

Fixes #1127